// src/app/interfaces/appSettings-ineterface.ts
export interface AppSettings {
    companyName?: string;
    companyContact?: string;
    companyAddress?: string;
    resetFormOnSave?: boolean;
    ccustomReciptMessage?: string;
    staticItems?: boolean;
    counterUser?: string;
    salesTaxPercent?: number;
    notificationDuraton?: number;
}